+++ apps/storybook/src/playground/routes.tsx
@@ -25,6 +25,7 @@ import PositionsPage from "./pages/portfolio/positions/page";
 import SettingsPage from "./pages/portfolio/setting/page";
 import AffiliatePage from "./pages/rewards/affiliate/page";
 import TradingRewardsPage from "./pages/rewards/trading/page";
+import VaultsPage from "./pages/vaults/page";
 import { getSymbol } from "./storage";
 
 const AppRoute = () => {
@@ -105,6 +106,10 @@ const AppRoute = () => {
     {
       path: "markets",
       element: <MarketsPage />,
+    },
+    {
+      path: "vaults",
+      element: <VaultsPage />,
     },
     {
       path: "leaderboard",