+++ packages/core/src/index.ts
@@ -21,6 +21,7 @@ export {
   generateAddOrderlyKeyMessage,
   generateRegisterAccountMessage,
   generateSettleMessage,
+  generateDexRequestMessage,
 } from "./helper";
 
 export {
@@ -49,6 +50,7 @@ export type {
   WithdrawInputs,
   SettleInputs,
   AddOrderlyKeyInputs,
+  DexRequestInputs,
 } from "./wallet/walletAdapter";
 export { BaseWalletAdapter } from "./wallet/baseWalletAdapter";
 export type { Ed25519Keypair } from "./types";