+++ packages/ui-transfer/src/components/depositAndWithdraw/plugin.tsx
@@ -1,0 +1,13 @@
+import { ExtensionPositionEnum, ExtensionSlot } from "@orderly.network/ui";
+import { DepositFormWidget } from "../depositForm";
+import { DepositFormWidgetProps } from "../depositForm/depositForm.widget";
+
+export const DepositSlot = (props: DepositFormWidgetProps) => {
+  return (
+    <ExtensionSlot
+      position={ExtensionPositionEnum.DepositForm}
+      defaultWidget={DepositFormWidget}
+      {...props}
+    />
+  );
+};