+++ packages/vaults/src/components/all-vaults/all-vaults.widget.tsx
@@ -1,0 +1,9 @@
+import { FC } from "react";
+import { useVaultInfoState } from "../../store/vaultsStore";
+import { AllVaultsDesktop } from "./all-vaults.desktop";
+
+export const AllVaultsWidget: FC = () => {
+  const vaultsInfo = useVaultInfoState().data;
+
+  return <AllVaultsDesktop vaults={vaultsInfo} />;
+};