+++ packages/vaults/src/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./all-vaults";
+export * from "./vault-card";
+export * from "./vaults-header";
+export * from "./vaults-introduction";
+export * from "./vault-operation";