+++ packages/vaults/src/components/vault-card/index.ts
@@ -1,0 +1,4 @@
+export * from "./vaultCard.widget";
+export * from "./vaultCard.script";
+export * from "./vaultCard.ui";
+export * from "./constants";