+++ packages/vaults/src/components/vault-operation/latest-deposit/latest-deposit.script.ts
@@ -1,0 +1,14 @@
+import { OperationType } from "../../../types/vault";
+import { useLatestOperationScript } from "../depositAndWithdraw/latest.script";
+
+export const useLatestDepositScript = ({ vaultId }: { vaultId: string }) => {
+  const { latestOperation } = useLatestOperationScript({
+    type: OperationType.DEPOSIT,
+    vaultId,
+  });
+  return {
+    latestOperation,
+  };
+};
+
+export type LatestDepositScript = ReturnType<typeof useLatestDepositScript>;