+++ packages/onboard/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/onboard
 
+## 2.6.0
+
+### Minor Changes
+
+- publish
+- c4b5fbc: publish
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [c4b5fbc]
+  - @orderly.network/hooks@2.6.0
+  - @orderly.network/types@2.6.0
+
+## 2.6.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/hooks@2.6.0-alpha.0
+  - @orderly.network/types@2.6.0-alpha.0
+
 ## 2.5.3
 
 ### Patch Changes