+++ packages/trading-view/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/trading-view
 
+## 2.6.0
+
+### Minor Changes
+
+- publish
+- c4b5fbc: publish
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [c4b5fbc]
+  - @orderly.network/core@2.6.0
+  - @orderly.network/hooks@2.6.0
+  - @orderly.network/net@2.6.0
+  - @orderly.network/types@2.6.0
+  - @orderly.network/utils@2.6.0
+
+## 2.6.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/core@2.6.0-alpha.0
+  - @orderly.network/hooks@2.6.0-alpha.0
+  - @orderly.network/net@2.6.0-alpha.0
+  - @orderly.network/types@2.6.0-alpha.0
+  - @orderly.network/utils@2.6.0-alpha.0
+
 ## 2.5.3
 
 ### Patch Changes